Agregación de nodos + su visual

+++ src/main/java/project/logic/StopNode.java
@@ -5,10 +5,10 @@ public class StopNode {
 
     private String idNodo;
     private String nombre;
-    private int posX;
-    private int posY;
+    private double posX;
+    private double posY;
 
-	public StopNode(String idNodo, String nombre, int posX, int posY) {
+	public StopNode(String idNodo, String nombre, double posX, double posY) {
         super();
         this.idNodo = idNodo;
         this.nombre = nombre;
@@ -32,22 +32,22 @@ public class StopNode {
     }
 
 
-    public int getPosX() {
+    public double getPosX() {
         return posX;
     }
 
 
-    public void setPosX(int posX) {
+    public void setPosX(double posX) {
         this.posX = posX;
     }
 
 
-    public int getPosY() {
+    public double getPosY() {
         return posY;
     }
 
 
-    public void setPosY(int posY) {
+    public void setPosY(double posY) {
         this.posY = posY;
     }
